dashboard/requirements.txt
streamlit
streamlit-folium
altair
boto3
python-dotenv
altair
vega_datasets
psycopg2-binary
st-pages
pandas
geopandas
pytest

